# [libc++][format] Improves escaping performance.