… libcxx/include/__format/escaped_output_table.h